Make current HTTP tracing work
Avoid introducing a Uploader.trace API to provide a lightweight client. When an error occurs, the stack trace would have create another level of Uploader.trace or whole stack of exceptions if an exception happened in our library. Rewrite them by Tracer enter/leaver primitives. Puting leave call in an ensure block, ensures the tracer is balanced (one enter call, one leave call).

For VCR and WebMock, finds the current span we are in and append a stubbed attribute to the span detail hash to distinguish andavoid creating a duplicate, incorrect trace.

Co-Authored-By: Matthew Draper <matthew@trebex.net>