Avoid activesupport when generating debugging timestamp
I believe DateTime.current is an activesupport-ism.

This gem does depend on activesupport so in theory we could use
DateTime.current, however when I was running v0.4.0 with my non-rails
rspec project I was getting:

   NameError: uninitialized constant RSpec::Buildkite::Analytics::Reporter::DateTime
   Did you mean?  DateAndTime
     /.../rspec-buildkite-analytics/lib/rspec/buildkite/analytics/reporter.rb:36:in `dump_summary'

I think we could fix this by adding an explicit `require "datetime"` to
the file. However if we're going to do that, we may as well stick to the
stdlib and use plain `Time.now` when generating a timestamp.