Better failure_reason and failure_expanded for Playwright reporter
This fixes a couple of issues:

1. When running tests with retries enabled, the failures before the final failure do not have `stack` or `snippet` attributes on `testResult.error` (both coming back undefined).  They _do_ have a `message` attribute, however.  Note _singular_ `testResult.error`.

2. However, `testResult.errors` (plural) _do_ contain stacks and snippets, but not necessarily on the first one returned by `error`.

So what this change does is:

1. For the `failure_reason`, return the first line of `testResult.error.message`.  This seems to always exist, and be the short-form message you'd expect here.

2. For `failure_expanded`, loop through all of the `testResult.errors` (plural), and append `stack` + `snippet` if they exist, and `message` otherwise.  Note that `stack` seems to always include `message`, so only one or the other is necessary.

This should _actually_ fix the `Error in reporter TypeError: Cannot read properties of undefined (reading 'split')` error reported in https://github.com/buildkite/test-collector-javascript/pull/78, which was actually because `testResult.error.stack` was often `undefined`.  (That PR is still worthwhile, since we should be properly including dependencies instead of just implicitly including them, but it didn't actually fix the error.  This should!)