Handle updated Upload API response, more permissive, better errors
The response body of the upload API is undocumented and not a
public/stable API. It was changed on the server-side, including the
removal of fields which this Swift collector considered mandatory.

This patch updates the response parsing to accept the new values from
the response (`upload_id`, `upload_url`) and to be more permissive about
future changes (handle these fields no longer existing).

We also now produce better error/warning/debug messages relating to
upload success/failure, including the addition of the HTTP response
status, which required some refactoring to use HTTPURLResponse.