rspec_example_trace_helpers: fix stubbed Hash initialization
Previously, `example.metadata` would return a Hash which, for any
unknown key, would return _another_ hash. I think this was not the
intention, but it didn't lead to any errors, probably because both Hash
and Array implement `#empty?`.

In Ruby 3.4, the syntax resulted in an error:

        unknown keyword: :shared_group_inclusion_backtrace

I believe the intention was for `example.metadata` to return the Hash
that contains the `shared_group_inclusion_backtrace` key, beginning as
an empty array.

The usage looks like this:

        lib/buildkite/test_collector/rspec_plugin/trace.rb
        63:      !example.metadata[:shared_group_inclusion_backtrace].empty?
        67:      example.metadata[:shared_group_inclusion_backtrace].last.inclusion_location