Add `location_prefix` option
Allows a file name prefix to be set which will be prepended to
`file_name` and `location` in the test results. Intended for use in
monorepos where the test suite lives in a sub directory of the project.

The change also DRYs some duplicate behaviour by extracting a `Trace`
superclass.